backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
supabase==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
groq>=0.5.0
pandas==2.1.3
pyarrow==14.0.1
aiofiles==23.2.1
sentence-transformers==2.2.2
torch==2.2.2+cpu
transformers==4.35.2
huggingface-hub==0.20.3
--extra-index-url https://download.pytorch.org/whl/cpu
